Restructured code:
